refactor(components): split components into each file
