+++ src/components/search/SearchBar.tsx
@@ -9,6 +9,7 @@ import { isHttpException, isServerActionError } from "../../server-action/backen
 import toast from "react-hot-toast";
 import { getInsideDoubleQuotes, parseKeyValue } from "./util"; 
 import { IPaginationResult } from "../../server-action/backend/common.dto";
+import { AutocompleteList } from "./AutoCompleteList";
 
 
 const baseSuggestion = Object.values(INPUT_KEY).map(str => str+':');
@@ -316,8 +317,9 @@ export function SearchBar({ onSearch, defaultValue,inputRef }: SearchBarProps) {
   }
 
 
-  const changeHandler = (e : React.ChangeEvent<HTMLInputElement>)=>{
-    console.log('changeHandler : ',e.target.value);
+
+
+  const changeHandler = ( e: React.ChangeEvent<HTMLInputElement>) => {
     const text = e.target.value;
 
     // 커서의 위치에 있는 문자 기준은 커서의 왼쪽 문자를 기준으로 한다.
@@ -385,55 +387,5 @@ export function SearchBar({ onSearch, defaultValue,inputRef }: SearchBarProps) {
           />
         )}
     </div>
-  );
-}
-
-
-interface AutocompleteListProps {
-  suggestions: string[];
-  onSelect: (value: string) => void;
-  query : string;
-  highlightedIndex : number,
-  setHighlightedIndex : (index : number) => void;
-}
-
-export function AutocompleteList({
-  suggestions,
-  onSelect,
-  query,
-  highlightedIndex,
-  setHighlightedIndex ,
-}: AutocompleteListProps) {
-  if (suggestions.length === 0) return null;
-
-  const highlightMatch = (text: string, query: string) => {
-    const i = text.toLowerCase().indexOf(query.toLowerCase());
-    if (i === -1) return text;
-    return (
-      <>
-        {text.slice(0, i)}
-        <span className="font-semibold text-red-400">{text.slice(i, i + query.length)}</span>
-        {text.slice(i + query.length)}
-      </>
-    );
-  };
-
-  return (
-    <ul
-      className={`absolute z-10 w-full mt-1 bg-gray-700 rounded-md shadow-lg max-h-60 overflow-y-auto `}
-    >
-      {suggestions.map((suggestion, index) => (
-        <li
-          key={index}
-          onClick={() => onSelect(suggestion)}
-          onMouseEnter={() => setHighlightedIndex(index)}
-          className={`px-4 py-2 cursor-pointer text-white ${
-            index === highlightedIndex ? 'bg-gray-600' : 'hover:bg-gray-600'
-          }`}
-        >
-          {highlightMatch(suggestion,query)}
-        </li>
-      ))}
-    </ul>
   );
 }
